:sparkles: refactor: 중복되는 클래스 삭제 및 예외 처리 수정
